+++ docs/whats-new.md
@@ -4,7 +4,7 @@
 
 ### Fix
 - robustly retrieving bbox in `dfmt.meshkernel_get_bbox()` in [#1292](https://github.com/Deltares/dfm_tools/pull/1292)
-
+- simpler parsing of uhslc geojson after pyogrio update in [#1294](https://github.com/Deltares/dfm_tools/pull/1294)
 
 ## 0.42.0 (2025-11-03)